moved MultiTouchTapDetector
