+++ packages/flame/lib/input.dart
@@ -6,6 +6,7 @@ export 'src/components/input/hud_button_component.dart';
 export 'src/components/input/hud_margin_component.dart';
 export 'src/components/input/joystick_component.dart';
 export 'src/components/input/sprite_button_component.dart';
+export 'src/events/game_mixins/multi_touch_tap_detector.dart';
 export 'src/extensions/vector2.dart';
 export 'src/game/mixins/keyboard.dart';
 export 'src/gestures/detectors.dart';
