+++ packages/flame/lib/src/gestures/detectors.dart
@@ -4,38 +4,6 @@ import '../events/multi_tap_listener.dart';
 import '../game/mixins/game.dart';
 import 'events.dart';
 
-mixin MultiTouchTapDetector on Game implements MultiTapListener {
-  void onTap(int pointerId) {}
-  void onTapCancel(int pointerId) {}
-  void onTapDown(int pointerId, TapDownInfo info) {}
-  void onTapUp(int pointerId, TapUpInfo info) {}
-  void onLongTapDown(int pointerId, TapDownInfo info) {}
-
-  @override
-  double get longTapDelay => 0.300;
-
-  @override
-  void handleTap(int pointerId) => onTap(pointerId);
-
-  @override
-  void handleTapCancel(int pointerId) => onTapCancel(pointerId);
-
-  @override
-  void handleTapDown(int pointerId, TapDownDetails details) {
-    onTapDown(pointerId, TapDownInfo.fromDetails(this, details));
-  }
-
-  @override
-  void handleTapUp(int pointerId, TapUpDetails details) {
-    onTapUp(pointerId, TapUpInfo.fromDetails(this, details));
-  }
-
-  @override
-  void handleLongTapDown(int pointerId, TapDownDetails details) {
-    onLongTapDown(pointerId, TapDownInfo.fromDetails(this, details));
-  }
-}
-
 mixin MultiTouchDragDetector on Game {
   void onDragStart(int pointerId, DragStartInfo info) {}
   void onDragUpdate(int pointerId, DragUpdateInfo info) {}
